tidy up Config things and move related bits into one place
