+++ subxt/src/config/extrinsic_params.rs
@@ -2,6 +2,11 @@
 // This file is dual-licensed as Apache-2.0 or GPL-3.0.
 // see LICENSE for license details.
 
+//! This module contains a trait which controls the parameters that can must
+//! be provided in order to successfully construct an extrinsic. A basic
+//! implementation of the trait is provided ([`BaseExtrinsicParams`]) which is
+//! used by the provided Substrate and Polkadot configuration.
+
 use crate::{
     utils::Encoded,
     Config,
@@ -47,22 +52,6 @@ pub trait ExtrinsicParams<Index, Hash>: Debug + 'static {
     /// signing it, meaning the client and node must agree on their values.
     fn encode_additional_to(&self, v: &mut Vec<u8>);
 }
-
-/// A struct representing the signed extra and additional parameters required
-/// to construct a transaction for the default substrate node.
-pub type SubstrateExtrinsicParams<T> = BaseExtrinsicParams<T, AssetTip>;
-
-/// A builder which leads to [`SubstrateExtrinsicParams`] being constructed.
-/// This is what you provide to methods like `sign_and_submit()`.
-pub type SubstrateExtrinsicParamsBuilder<T> = BaseExtrinsicParamsBuilder<T, AssetTip>;
-
-/// A struct representing the signed extra and additional parameters required
-/// to construct a transaction for a polkadot node.
-pub type PolkadotExtrinsicParams<T> = BaseExtrinsicParams<T, PlainTip>;
-
-/// A builder which leads to [`PolkadotExtrinsicParams`] being constructed.
-/// This is what you provide to methods like `sign_and_submit()`.
-pub type PolkadotExtrinsicParamsBuilder<T> = BaseExtrinsicParamsBuilder<T, PlainTip>;
 
 /// An implementation of [`ExtrinsicParams`] that is suitable for constructing
 /// extrinsics that can be sent to a node with the same signed extra and additional
@@ -182,56 +171,5 @@ impl<T: Config, Tip: Debug + Encode + 'static> ExtrinsicParams<T::Index, T::Hash
             self.mortality_checkpoint,
         )
             .encode_to(v);
-    }
-}
-
-/// A tip payment.
-#[derive(Copy, Clone, Debug, Default, Encode)]
-pub struct PlainTip {
-    #[codec(compact)]
-    tip: u128,
-}
-
-impl PlainTip {
-    /// Create a new tip of the amount provided.
-    pub fn new(amount: u128) -> Self {
-        PlainTip { tip: amount }
-    }
-}
-
-impl From<u128> for PlainTip {
-    fn from(n: u128) -> Self {
-        PlainTip::new(n)
-    }
-}
-
-/// A tip payment made in the form of a specific asset.
-#[derive(Copy, Clone, Debug, Default, Encode)]
-pub struct AssetTip {
-    #[codec(compact)]
-    tip: u128,
-    asset: Option<u32>,
-}
-
-impl AssetTip {
-    /// Create a new tip of the amount provided.
-    pub fn new(amount: u128) -> Self {
-        AssetTip {
-            tip: amount,
-            asset: None,
-        }
-    }
-
-    /// Designate the tip as being of a particular asset class.
-    /// If this is not set, then the native currency is used.
-    pub fn of_asset(mut self, asset: u32) -> Self {
-        self.asset = Some(asset);
-        self
-    }
-}
-
-impl From<u128> for AssetTip {
-    fn from(n: u128) -> Self {
-        AssetTip::new(n)
     }
 }
