+++ subxt/src/tx/tx_progress.rs
@@ -370,7 +370,7 @@ impl<T: Config, C: OnlineClientT<T>> TxInBlock<T, C> {
         let extrinsic_idx = block.block.extrinsics
             .iter()
             .position(|ext| {
-                use crate::utils::hasher::Hasher;
+                use crate::config::Hasher;
                 let hash = T::Hashing::hash_of(&ext.0);
                 hash == self.ext_hash
             })
